Refactor Services, Repositories, and Add Unit Tests

+++ app/Services/CommentService.php
@@ -4,16 +4,29 @@ namespace App\Services;
 
 use App\Http\Repositories\CommentRepository;
 use App\Http\Resources\Comment\CommentResource;
+use Illuminate\Contracts\Queue\EntityNotFoundException;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
 class CommentService {
 
     public function __construct(private readonly CommentRepository $repository) {}
 
+    public function index(): AnonymousResourceCollection {
+        return CommentResource::collection($this->repository->index());
+    }
+
     public function show(int $id): CommentResource {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Comment', $id);
+        }
         return new CommentResource($this->repository->show($id));
     }
 
     public function update(int $id, mixed $data): CommentResource {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Comment', $id);
+        }
+
         $comment = $this->repository->show($id);
 
         return new CommentResource($this->repository->update($comment, $data));
@@ -24,6 +37,9 @@ class CommentService {
     }
 
     public function destroy(int $id): void {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Comment', $id);
+        }
         $this->repository->destroy($id);
     }
 }
